drivers/perf: riscv: Remove the warning from stop function
The warning messsage was initially added just to indicate that counter
stop function is being called while the event is already stopped.

However, we update the state to stopped case now in an overflow handler
after stopping the counter. If there is another child overflow handler
is registered (e.g kvm) it may call stop again which will trigger the
warning.

Signed-off-by: Atish Patra <atishp@rivosinc.com>