drivers/irqchip: Add GTMB driver for GSI based ACPI devices
The GTMB driver acts as the GSI-to-MSI bridge for any MSI capable device
that uses GSI infrastructure in ACPI. One of the example is the GED device
which can be used on platforms without a centralized IRQ-to-MSI bridge.
The GTMB device will become the parent IRQ domain for the GED device
interrupt.

Currently, it only implements the generic registers and System
memory as the address space type. Currently, it doesn't support
other address space type but can be easily extended to support it.

Signed-off-by: Atish Patra <atishp@rivosinc.com>