KVM: riscv: selftests: Change command line option
The PMU test commandline option takes an argument
to disable a certain test. The initial objective is that
use probably want just to run all the test most of the time.
However, it is found that running a single test is more useful
instead. Especially, the overflow test has been helpful to
validate PMU virtualizaiton changes.

Switching the command line option to run a single test instead
of disabling a single test also allows to provide additional
test specific arguments to the test.

Signed-off-by: Atish Patra <atishp@rivosinc.com>